refactor: extract home screen components to standalone files
This commit refactors the `HomeScreen.kt` by extracting several UI components into a dedicated `components` package. This improves code maintainability, readability, and allows for better component reuse across the KoColor feature.

- **Component Extraction**:
    - Created `BoutiqueCard.kt`: Moved the `BoutiqueCard` composable, which features an elevated card with an image background and a white overlay for boutique information.
    - Created `QuickActions.kt`: Moved `QuickActions`, `QuickActionCard`, and the `QuickActionUiState` data class. This component handles the AI visual analysis and product scanner entry points.
    - Created `SectionTitle.kt`: Moved the `SectionTitle` composable and `SectionTitleUiState` to provide a standardized header style for screen sections.
- **Code Cleanup**:
    - Removed redundant component definitions from `HomeScreen.kt`.
    - Updated imports in `HomeScreen.kt` and `WellnessInsightsSection.kt` to reference the new component locations.
    - Added default parameter values to event and navigation callbacks in the extracted components for better call-site flexibility.
